+++ lib/ts/components/componentOverride/componentOverrideContext.tsx
@@ -2,10 +2,10 @@ import React from "react";
 
 import type { ComponentOverride } from "./componentOverride";
 
-type ContextType<T> =
-    | {
-          [K in keyof T]?: ComponentOverride<any>;
-      }
-    | "IS_DEFAULT";
+export type GenericComponentOverrideMap<T> = {
+    [K in keyof T]?: ComponentOverride<any>;
+};
+
+type ContextType<T> = GenericComponentOverrideMap<T> | "IS_DEFAULT";
 
 export const ComponentOverrideContext = React.createContext<ContextType<any>>("IS_DEFAULT");