+++ lib/ts/recipe/emailpassword/componentOverrideContext.tsx
@@ -2,10 +2,6 @@ import { createGenericComponentsOverrideContext } from "../../components/compone
 
 import type { ComponentOverrideMap } from "./types";
 
-const [useContext, Provider, Consumer] = createGenericComponentsOverrideContext<ComponentOverrideMap>();
+const [useContext, Provider] = createGenericComponentsOverrideContext<ComponentOverrideMap>();
 
-export {
-    useContext as useRecipeComponentOverrideContext,
-    Provider as RecipeComponentsOverrideContextProvider,
-    Consumer as RecipeComponentsOverrideContextConsumer,
-};
+export { useContext as useRecipeComponentOverrideContext, Provider as RecipeComponentsOverrideContextProvider };