+++ lib/ts/recipe/emailverification/components/features/emailVerification/index.tsx
@@ -26,21 +26,21 @@ import { useUserContext } from "../../../../../usercontext";
 import { clearQueryParams, getQueryParams, useOnMountAPICall } from "../../../../../utils";
 import { SessionContext } from "../../../../session";
 import Session from "../../../../session/recipe";
-import { useRecipeComponentOverrideContext } from "../../../componentOverrideContext";
 import EmailVerificationTheme from "../../themes/emailVerification";
 import { defaultTranslationsEmailVerification } from "../../themes/translations";
 
 import type { FeatureBaseProps } from "../../../../../types";
 import type Recipe from "../../../recipe";
+import type { ComponentOverrideMap } from "../../../types";
 import type { RecipeInterface } from "supertokens-web-js/recipe/emailverification";
 
-type Prop = FeatureBaseProps & { recipe: Recipe; userContext?: any };
+type Prop = FeatureBaseProps & { recipe: Recipe; userContext?: any; useComponentOverrides: () => ComponentOverrideMap };
 
 export const EmailVerification: React.FC<Prop> = (props) => {
     const sessionContext = useContext(SessionContext);
     const [status, setStatus] = useState("LOADING");
     const userContext = useUserContext();
-    const recipeComponentOverrides = useRecipeComponentOverrideContext();
+    const recipeComponentOverrides = props.useComponentOverrides();
 
     const redirectToAuthWithHistory = useCallback(async () => {
         await redirectToAuth({ redirectBack: false, history: props.history });