+++ lib/ts/recipe/emailverification/recipe.tsx
@@ -30,6 +30,7 @@ import { matchRecipeIdUsingQueryParams, saveInvalidClaimRedirectPathInContext } 
 import RecipeModule from "../recipeModule";
 import { SessionAuth } from "../session";
 
+import { useRecipeComponentOverrideContext } from "./componentOverrideContext";
 import { default as EmailVerificationFeature } from "./components/features/emailVerification";
 import { DEFAULT_VERIFY_EMAIL_PATH } from "./constants";
 import RecipeImplementation from "./recipeImplementation";
@@ -43,6 +44,7 @@ import type {
     OnHandleEventContext,
     PreAndPostAPIHookAction,
 } from "./types";
+import type { GenericComponentOverrideMap } from "../../components/componentOverride/componentOverrideContext";
 import type { RecipeFeatureComponentMap } from "../../types";
 import type { CreateRecipeFunction, NormalisedAppInfo } from "../../types";
 import type { RecipeInterface } from "supertokens-web-js/recipe/emailverification";
@@ -122,7 +124,9 @@ export default class EmailVerification extends RecipeModule<
         return EmailVerification.instance;
     }
 
-    getFeatures = (): RecipeFeatureComponentMap => {
+    getFeatures = (
+        useComponentOverrides: () => GenericComponentOverrideMap<any> = useRecipeComponentOverrideContext
+    ): RecipeFeatureComponentMap => {
         const features: RecipeFeatureComponentMap = {};
         if (this.config.disableDefaultUI !== true) {
             const normalisedFullPath = this.config.appInfo.websiteBasePath.appendPath(
@@ -130,14 +134,18 @@ export default class EmailVerification extends RecipeModule<
             );
             features[normalisedFullPath.getAsStringDangerous()] = {
                 matches: matchRecipeIdUsingQueryParams(this.config.recipeId),
-                component: (props: any) => this.getFeatureComponent("emailverification", props),
+                component: (props: any) => this.getFeatureComponent("emailverification", props, useComponentOverrides),
             };
         }
         return features;
     };
 
-    // eslint-disable-next-line @typescript-eslint/no-unused-vars
-    getFeatureComponent = (_: "emailverification", props: any): JSX.Element => {
+    getFeatureComponent = (
+        // eslint-disable-next-line @typescript-eslint/no-unused-vars
+        _: "emailverification",
+        props: any,
+        useComponentOverrides: () => GenericComponentOverrideMap<any> = useRecipeComponentOverrideContext
+    ): JSX.Element => {
         return (
             <UserContextWrapper userContext={props.userContext}>
                 <SessionAuth requireAuth={false} overrideGlobalClaimValidators={() => []}>
@@ -158,6 +166,7 @@ export default class EmailVerification extends RecipeModule<
                             return (
                                 <EmailVerificationFeature
                                     recipe={this}
+                                    useComponentOverrides={useComponentOverrides}
                                     {...{
                                         ...props,
                                         // We do this to make sure it does not add another provider