+++ lib/ts/recipe/recipeModule/index.ts
@@ -17,6 +17,7 @@ import SuperTokens from "../../superTokens";
 import { appendQueryParamsToURL } from "../../utils";
 
 import type { NormalisedConfig } from "./types";
+import type { GenericComponentOverrideMap } from "../../components/componentOverride/componentOverrideContext";
 import type { RecipeFeatureComponentMap } from "../../types";
 
 /*
@@ -64,7 +65,11 @@ export default abstract class RecipeModule<
         throw new Error("getDefaultRedirectionURL is not implemented.");
     }
 
-    abstract getFeatures(): RecipeFeatureComponentMap;
+    abstract getFeatures(useComponentOverrides?: () => GenericComponentOverrideMap<any>): RecipeFeatureComponentMap;
 
-    abstract getFeatureComponent(componentName: string, props: any): JSX.Element;
+    abstract getFeatureComponent(
+        componentName: string,
+        props: any,
+        useComponentOverrides?: () => GenericComponentOverrideMap<any>
+    ): JSX.Element;
 }